backend/pyproject.toml
[project]
name = "neuraldamage-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.12.1",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
    "wsproto>=1.3.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
